feat: refactor linter into smaller modules
