+++ conda_smithy/linter/license_exceptions.txt
