+++ conda_smithy/linter/licenses.txt
